move AppSelector to not talk to the store directly
